Added refresh function
